video: Moved EncodedFrame, DecodedFrame and FrameDependency from core to video
