+++ core/src/backend/video/software/mod.rs
@@ -1,12 +1,11 @@
 //! Pure software video decoding backend.
 
-use crate::backend::video::{
-    DecodedFrame, EncodedFrame, FrameDependency, VideoBackend, VideoStreamHandle,
-};
+use crate::backend::video::{VideoBackend, VideoStreamHandle};
 use generational_arena::Arena;
 use ruffle_render::backend::RenderBackend;
 use ruffle_render::bitmap::{Bitmap, BitmapFormat, BitmapHandle, BitmapInfo};
 use ruffle_video::error::Error;
+use ruffle_video::frame::{DecodedFrame, EncodedFrame, FrameDependency};
 use swf::{VideoCodec, VideoDeblocking};
 
 pub mod decoders;
