+++ video/src/lib.rs
@@ -1,1 +1,2 @@
 pub mod error;
+pub mod frame;
